requirements.txt
streamlit==1.43.2
langchain==0.0.285
pypdf==3.16.0
docx2txt==0.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.0.1
accelerate==0.23.0
langchain_community
pymupdf
python-dotenv